+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 
 - `PUT /changelist/{project}/{name}` no longer supports renaming a changelist
 
+### Fixed
+
+- Return 400 when new or renamed changelist will have same name as another existing one
+
 ## [0.1.0-beta.2] - 2023-12-02
 
 ### Added